# backend-node/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 宽窄·Orbit 业务后端 API
  description: |
    宽窄·Orbit 社区型 APP 业务后端 API 契约。

    ## 认证
    使用 Bearer Token 方式：`Authorization: Bearer <access_token>`

    ## 统一响应格式
    ```json
    { "success": true, "data": { ... } }
    ```
    错误时：
    ```json
    { "success": false, "error": { "code": "40001", "message": "..." }, "requestId": "req_xxx" }
    ```

    ## 错误码
    | Code | 含义 |
    |------|------|
    | 40001 | 请求参数错误 |
    | 40101 | 未登录或 Token 失效 |
    | 40301 | 无权限或被限制 |
    | 40401 | 资源不存在 |
    | 40901 | 状态冲突 / 重复操作 / 额度不足 |
    | 42901 | 请求频率过高 |
    | 50000 | 服务内部错误 |
    | 50401 | 上游超时 |

    ## 幂等
    写操作推荐携带 `Idempotency-Key` 请求头（UUID），相同 Key 的重复请求返回相同结果。

    ## 分页
    使用 page-based 分页：`?page=1&pageSize=20`，响应包含 `hasMore` 和 `nextPage`。
  version: 1.0.0
  contact:
    name: 宽窄·Orbit Backend Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.kuanzhai-orbit.com/api/v1
    description: Production
...
paths:
...
  /ritual/perform:
    post:
      tags: [ritual]
      summary: 创建仪式会话（起卦）
      operationId: performRitual
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RitualPerformRequest'
      responses:
        '200':
          description: 仪式会话已创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RitualSessionResponse'
  /ritual/session/{sessionId}:
    get:
      tags: [ritual]
      summary: 恢复仪式会话
      operationId: getRitualSession
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 会话详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RitualSessionResponse'
  /ritual/session/{sessionId}/preview:
    get:
      tags: [ritual]
      summary: 获取解读预览（首次体验）
      operationId: getRitualPreview
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 预览内容
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InterpretationCard'
  /ritual/session/{sessionId}/full-read:
    get:
      tags: [ritual]
      summary: 获取完整解读
      operationId: getFullInterpretation
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 完整解读内容
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InterpretationCard'
  /ritual/session/{sessionId}/continue:
    post:
      tags: [ritual]
      summary: 追问
      operationId: continueRitual
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: 追问回复
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FollowupMessage'
  /ritual/session/{sessionId}/chat-history:
    get:
      tags: [ritual]
      summary: 获取对话历史
      operationId: getChatHistory
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 对话历史
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          messages:
                            type: array
                            items:
                              $ref: '#/components/schemas/FollowupMessage'
  /ritual/session/{sessionId}/interpretation/stream:
    get:
      tags: [ritual]
      summary: SSE 流式解读
      operationId: streamInterpretation
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  事件格式：
                  - `event: chunk\ndata: {"delta":"..."}\n\n`
                  - `event: done\ndata: {"sessionId":"..."}\n\n`
                  - `event: error\ndata: {"code":"...","message":"..."}\n\n`
  /ritual/session/{sessionId}/followup/stream:
    get:
      tags: [ritual]
      summary: SSE 流式追问回复
      operationId: streamFollowup
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
  /ritual/user/{userId}/completion-today:
    get:
      tags: [ritual]
      summary: 获取今日完成状态
      operationId: getCompletionToday
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 今日完成状态
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          completed:
                            type: boolean
                          dateKey:
                            type: string

  # ─────────────── Community ───────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    RitualPerformRequest:
      type: object
      required: [question, tag, lines, movingLines]
      properties:
        question:
          type: string
          maxLength: 500
        tag:
          type: string
          enum: [relationship, career, emotion, choice, other]
        lines:
          type: array
          items:
            type: integer
            enum: [0, 1]
          minItems: 6
          maxItems: 6
        movingLines:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 5
    RitualSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        question:
          type: string
        tag:
          type: string
        pattern:
          type: object
          properties:
            lines:
              type: array
              items:
                type: integer
            movingLines:
              type: array
              items:
                type: integer
        status:
          type: string
        card:
          $ref: '#/components/schemas/InterpretationCard'
        createdAt:
          type: string
          format: date-time
    InterpretationCard:
      type: object
      properties:
        summary:
          type: string
        body:
          type: string
        followupDirections:
          type: array
          items:
            type: string
        needsClarification:
          type: boolean
        microActions:
          type: array
          items:
            type: string
        focusPoints:
          type: array
          items:
            type: string
        afterglow:
          type: string
    FollowupMessage:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [question, answer]
        content:
          type: string
        suggestedDirections:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...